+++ frontend/site-registry/src/app/features/map/SiteMarkers/SiteMarkers.tsx
